Draws Properly.

+++ CPLUSPLUS/InitGL.h
@@ -3,8 +3,8 @@
 
 #include "OpenGLHeaders.h"
 
-const float WINDOW_WIDTH = 800.0;
-const float WINDOW_HEIGHT = 600.0;
+const int WINDOW_WIDTH = 800;
+const int WINDOW_HEIGHT = 600;
 const int SCREEN_FPS = 60;
 
 void Initialize_glDisplay();
